HEEDLS-541 refactor, post-rebase fixes
